move closeDocument to extension.ts
